Refactor NE resource loader to a separate class.
